feat: externalize cache building and simplify IVoxelMap
