+++ src/lib/index.ts
@@ -1,3 +1,3 @@
 export { ELogLevel, setVerbosity } from './helpers/logger';
-export type { IVoxel, IVoxelMap, IVoxelMaterial } from './terrain/i-voxel-map';
+export type { IVoxelMap, IVoxelMaterial, ILocalMapData } from './terrain/i-voxel-map';
 export { EPatchComputingMode, Terrain } from './terrain/terrain';
